I added work on over fitting and under fitting of decision trees and intro the the random forest
Overfitting and underfitting are about trying to find the sweet spot where the model makes the most accurate predictions. This is done by running the model at different heights of the decision tree and selecting the height at which the model has the least MSE.
The random forest uses many trees, and it makes a prediction by averaging the predictions of each component tree. It generally has much better predictive accuracy than a single decision tree and it works well with default parameters. If you keep modeling, you can learn more models with even better performance, but many of those are sensitive to getting the right parameters.